move to doubles for floats

+++ audio/src/localization.rs
@@ -2,13 +2,13 @@
 
 use nalgebra::{Const, OMatrix, OVector};
 
-const SPEED_OF_SOUND: f32 = 343.260;
+const SPEED_OF_SOUND: f64 = 343.260;
 
 #[must_use]
 pub fn compute_direction<const DIM: usize, const N_MICS: usize>(
-    positions: &OMatrix<f32, Const<DIM>, Const<N_MICS>>,
-    times: &OVector<f32, Const<N_MICS>>,
-) -> OVector<f32, Const<DIM>> {
+    positions: &OMatrix<f64, Const<DIM>, Const<N_MICS>>,
+    times: &OVector<f64, Const<N_MICS>>,
+) -> OVector<f64, Const<DIM>> {
     let mut adjusted_distances = times.clone_owned();
 
     let min_time_idx = (0..times.len())
@@ -47,12 +47,12 @@ mod tests {
 
     #[test]
     fn it_works() {
-        let mic_points = SMatrix::<f32, 2, 8>::from_row_slice(&[
+        let mic_points = SMatrix::<f64, 2, 8>::from_row_slice(&[
             -0.09, -0.14, -0.105, -0.01, 0.09, 0.12, 0.085, 0.0, // x positions
             0.095, 0.015, -0.06, -0.115, -0.085, 0.0, 0.105, 0.16, // y positions
         ]);
 
-        let mic_times = SVector::<f32, 8>::from_row_slice(&[
+        let mic_times = SVector::<f64, 8>::from_row_slice(&[
             0.0004, 0.0005, 0.0007, 0.0008, 0.0007, 0.0005, 0.0004, 0.0001,
         ]);
